4.17

+++ controllers/blogs.js
@@ -1,32 +1,46 @@
 const express = require('express')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 const blogsRouter = express.Router()
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const body = request.body
+  const { title, author, url, likes } = request.body
 
-  if (!body.title || !body.url) {
+  if (!title || !url) {
     return response.status(400).json({ error: 'title or url missing' })
   }
 
+  const users = await User.find({})
+  const user = users[0] // Replace with actual logged-in user from token later
+
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
+    title,
+    author,
+    url,
+    likes: likes || 0,
+    user: user._id,
   })
 
   const savedBlog = await blog.save()
-  response.status(201).json(savedBlog)
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  const populatedBlog = await savedBlog.populate('user', { username: 1, name: 1 })
+  response.status(201).json(populatedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
